[Doc] Improve Capsule with Jinyang & Fix wrong tutorial level layout (#236)
* improve capsule tutorial with jinyang

* fix wrong layout of second-level tutorial

* delete transformer
