+++ docs/source/conf.py
@@ -195,8 +195,8 @@ intersphinx_mapping = {
 # sphinx gallery configurations
 from sphinx_gallery.sorting import FileNameSortKey
 
-examples_dirs = ['../../tutorials']  # path to find sources
-gallery_dirs = ['tutorials']  # path to generate docs
+examples_dirs = ['../../tutorials/basics','../../tutorials/models']  # path to find sources
+gallery_dirs = ['tutorials/basics','tutorials/models']  # path to generate docs
 reference_url = {
     'dgl' : None,
     'numpy': 'http://docs.scipy.org/doc/numpy/',
