--- tutorials/README.txt
@@ -1,4 +1,0 @@
-Basic Tutorials
-===============
-
-These tutorials conver the basics of DGL.
